backend/app/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
torch==2.4.1
transformers==4.30.0
pillow==10.2.0
requests==2.28.2
python-multipart==0.0.5
numpy==1.26.1
opencv-python==4.7.0.72